Added BuildingBlocks utility
This is a utility that allows for building blocks to be added far quicker in the future using a single line of code.

The first block added where the barn blocks. They are found exclusively in barns and can be used just like any other planks. They have a rougher, warren texture.